HARP-8263: Add mermaid diagrams to coresdk documentation.
See docs/@docs/README.md for more information.
