+++ docs/@docs/README.md
@@ -6,6 +6,16 @@ This directory contains markdown documents that will be integrated to typedoc ou
 
 1. Create a new subdirectory with the name that should be used to identify the document at the
    table of contents generated by typedoc.
-2. Create a file named 'README.md' in the subdirectory and add there the documentation.
-3. Add an empty ts file to force typedoc to process the subdirectory.
+2. Create a file named *README.md* in the subdirectory and add there the documentation.
+3. Add an dummy ts file (it must have a license header) to force typedoc to process the subdirectory.
+4. Optionally, add *.mmd* files containing mermaid diagrams. These will be rendered into svg files
+   which can be included as images in any markdown file. E.g.:
+    * A mermaid diagram is located at *@docs/dir1/diagram1.mmd*.
+    * The diagram can be added to a markdown file using the reference *media://dir1/diagram1.svg*.
+    * Since diagrams are large, it's better to add them as thumbnails linking to the full-sized image using HTML:
+   ```html
+    <a href="media://dir1/diagram1.svg">
+        <img src="media://dir1/diagram1.svg" alt="drawing1" title="Diagram" width="100"/>
+    </a>
+    ```
