+++ package.json
@@ -21,6 +21,7 @@
         "geckodriver": "^1.16.2",
         "glob": "^7.1.4",
         "hard-source-webpack-plugin": "^0.13.1",
+        "@ismarslomic/mermaid.cli": "^0.5.3",
         "mkpath": "^1.0.0",
         "mocha": "^6.1.4",
         "mocha-webdriver-runner": "^0.5.3",
@@ -58,7 +59,7 @@
         "run-rendering-tests": "./scripts/rendering-tests-reference-platform.sh",
         "save-reference-rendering-tests": "npx ts-node @here/harp-test-utils/lib/rendering/RenderingTestResultCli.ts save-reference",
         "approve-reference-rendering-tests": "npx ts-node @here/harp-test-utils/lib/rendering/RenderingTestResultCli.ts approve",
-        "typedoc": "ts-node ./scripts/doc-snippets.ts && typedoc --disableOutputCheck --options typedoc.json",
+        "typedoc": "ts-node ./scripts/setup-docs.ts && typedoc --disableOutputCheck --options typedoc.json",
         "tslint": "tslint --project tsconfig.json",
         "tslint:fix": "tslint --fix --project tsconfig.json",
         "prettier": "prettier -l \"**/*.ts\" \"**/*.tsx\" \"**/*.json\"",
